Analytics now over to retrofit
